Added copy folder support

+++ src/test/java/com/sony/ebs/octopus3/commons/file/FileUtilsTest.java
@@ -29,6 +29,10 @@ public class FileUtilsTest {
     Path filePath3 = Paths.get(basePath + "/file3.txt");
     Path filePath4 = Paths.get(basePath + "/d/file4.txt");
     Path filePath5 = Paths.get(basePath + "/d/e/file5.txt");
+    Path filePath6 = Paths.get(basePath + "/d");
+    Path filePath7 = Paths.get(basePath + "/d/file7.txt");
+    Path filePath8 = Paths.get(basePath + "/d/f/file8.txt");
+    Path filePath9 = Paths.get(basePath + "/g/h");
     Path zipPath = Paths.get(basePath.getParent() + "/a.zip");
 
     @Before
@@ -39,9 +43,13 @@ public class FileUtilsTest {
     @After
     public void doAfter() {
         resetFilePermissions();
-        FileUtils.delete(basePath);
+        if (basePath.toFile().exists()) {
+            FileUtils.delete(basePath);
+        }
         //Delete zip file in zip tests
-        FileUtils.delete(zipPath);
+        if (zipPath.toFile().exists()) {
+            FileUtils.delete(zipPath);
+        }
     }
 
     @Test
@@ -224,7 +232,7 @@ public class FileUtilsTest {
         assertTrue(result.getTracked().contains(filePath4));
         assertTrue(result.getTracked().contains(filePath5));
 
-        validateZip(zipPath, Arrays.asList("d/", "d/e/", "d/e/file5.txt",  "d/file4.txt", "file1.txt", "file2.txt"));
+        validateZip(zipPath, Arrays.asList("d/", "d/e/", "d/e/file5.txt", "d/file4.txt", "file1.txt", "file2.txt"));
     }
 
     @Test(expected = InstantiationException.class)
@@ -238,6 +246,26 @@ public class FileUtilsTest {
             // no need to expect reflection errors
             // we are interested in our own exceptions
         }
+    }
+
+    @Test
+    public void copyFolder() throws IOException {
+        FileUtils.writeFile(filePath7, "test".getBytes(), true, true);
+        FileUtils.writeFile(filePath8, "test".getBytes(), true, true);
+
+        FileUtils.copy(filePath6, filePath9);
+
+        assertTrue(Paths.get(filePath9 + "/file7.txt").toFile().exists());
+        assertTrue(Paths.get(filePath9 + "/f/file8.txt").toFile().exists());
+    }
+
+    @Test
+    public void copyFile() throws IOException {
+        FileUtils.writeFile(filePath7, "test".getBytes(), true, true);
+
+        FileUtils.copy(filePath7, filePath9);
+
+        assertTrue(Paths.get(filePath9 + "/file7.txt").toFile().exists());
     }
 
     private static void resetFilePermissions() {
